+++ irohad/CMakeLists.txt
@@ -13,7 +13,6 @@ add_subdirectory(network)
 add_subdirectory(model)
 add_subdirectory(simulator)
 add_subdirectory(synchronizer)
-add_subdirectory(multi_sig_transactions)
 add_subdirectory(pending_txs_storage)
 add_subdirectory(util)
 add_subdirectory(maintenance)
