+++ irohad/main/CMakeLists.txt
@@ -88,7 +88,6 @@ target_link_libraries(application
     simulator
     block_loader
     block_loader_service
-    mst_processor
     torii_service
     pending_txs_storage
     common
