+++ irohad/ordering/CMakeLists.txt
@@ -16,8 +16,6 @@ add_library(on_demand_ordering_service
 
 target_link_libraries(on_demand_ordering_service
     on_demand_common
-    mst_hash
-    mst_state
     shared_model_interfaces
     consensus_round
     logger
