+++ irohad/pending_txs_storage/CMakeLists.txt
@@ -8,7 +8,6 @@ add_library(pending_txs_storage
     )
 
 target_link_libraries(pending_txs_storage
-    mst_state
     shared_model_interfaces
     rxcpp
     )
