+++ irohad/pending_txs_storage/impl/pending_txs_storage_impl.cpp
@@ -9,7 +9,6 @@
 
 #include "ametsuchi/tx_presence_cache_utils.hpp"
 #include "interfaces/transaction.hpp"
-#include "multi_sig_transactions/state/mst_state.hpp"
 
 using iroha::PendingTransactionStorageImpl;
