+++ shared_model/interfaces/iroha_internal/transaction_batch.cpp
@@ -23,5 +23,12 @@ namespace shared_model {
         const std::shared_ptr<TransactionBatch> &right_tx) const {
       return left_tx->reducedHash() == right_tx->reducedHash();
     }
+
+     size_t BatchPointerHasher::operator()(
+          const std::shared_ptr<shared_model::interface::TransactionBatch> &a)
+          const {
+        return hasher_(a->reducedHash());
+      }
+
   }  // namespace interface
 }  // namespace shared_model
