+++ shared_model/interfaces/iroha_internal/transaction_batch.hpp
@@ -69,6 +69,13 @@ namespace shared_model {
       bool operator()(const std::shared_ptr<TransactionBatch> &left_tx,
                       const std::shared_ptr<TransactionBatch> &right_tx) const;
     };
+
+    struct BatchPointerHasher {
+      shared_model::crypto::Hash::Hasher hasher_;
+      size_t operator()(
+          const std::shared_ptr<shared_model::interface::TransactionBatch> &a)
+          const ;
+    };
   }  // namespace interface
 }  // namespace shared_model
